pyproject.toml
[tool.poetry]
name = "teste-python-backend"
version = "0.1.0"
description = ""
authors = ["Marcos R. Morato <marcos.nyrou@hotmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
environs = "^9.3.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^2.7.0"
psycopg2-binary = "^2.8.6"
Flask-JWT-Extended = "^4.1.0"
selenium = "^3.141.0"
geckodriver-autoinstaller = "^0.1.0"
pytest = "^6.2.3"
Faker = "^8.1.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flask-sqlalchemy-stubs = {git = "https://github.com/ssfdust/flask-sqlalchemy-stubs.git"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
